Fix icon picker crash: extract localStorage utils from React Compiler scope
The React Compiler (compilationMode: 'all') was injecting useMemoCache
into saveRecentIcon/getRecentIcons — plain functions defined in
IconPicker.tsx. When called from onClick, React threw "Invalid hook
call" and the icon selection silently failed.

Fix: move these functions to recent-icons.ts with 'use no memo'.
